Rewrite rsqrt with Arm64/NEON counterpart
rsqrt is the operation for floating reciprocal square root estimate.
In Arm64, we can implement it with NEON intrinsics. Since Armv8.2,
instruction "FRSQRTE" [1] is provided to calculate an approximate square
root for each vector element in the source SIMD and FP register.

With -O3, generated assembly on Apple Silicon M1:

[original]
        fsqrt   s0, s0
        fmov    s1, #1.00000000
        fdiv    s0, s1, s0

[neon]
        dup.4s  v0, v0[0]
        frsqrte.4s      v0, v0

[1] https://developer.arm.com/documentation/dui0801/k/A64-SIMD-Vector-Instructions/FRSQRTE--vector-